tazmania/relayer/requirements.txt
anyio==3.6.2
base58==2.1.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ed25519==1.5
fastapi==0.85.2
h11==0.14.0
idna==3.4
near-api-py==0.1.0
pydantic==1.10.2
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
tinydb==4.7.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
